+++ src/urlopen.h
@@ -34,5 +34,5 @@ int url_fclose(URL_FILE *file);
 int url_feof(URL_FILE *file);
 size_t url_fread(void *ptr, size_t size, size_t nmemb, URL_FILE *file);
 char *url_fgets(char *ptr, size_t size, URL_FILE *file);
-ssize_t url_getline(STRING *line, URL_FILE *file);
+int url_getline(STRING *line, URL_FILE *file);
 void url_rewind(URL_FILE *file);
